+++ core/roboconf-agent-monitoring/src/main/java/net/roboconf/agent/monitoring/internal/MonitoringTask.java
@@ -152,7 +152,7 @@ public class MonitoringTask extends TimerTask {
 				sb.setLength( 0 );
 			}
 
-			sb.append( expandString(s, params) + "\n" );
+			sb.append( Utils.expandTemplate(s, params) + "\n" );
 		}
 
 		addSectionIfNotEmpty(sections, sb.toString());
@@ -182,32 +182,7 @@ public class MonitoringTask extends TimerTask {
 		if(section.trim().length() > 0) sections.add(section);
 	}
 
-	/**
-	 * Expand a template, replacing each {{ param }} by the corresponding value.
-	 * Eg. "My name is {{ name }}" will result in "My name is Bond", provided that "params" contains "name=Bond".
-	 * @param s The template to expand
-	 * @param params The parameters to be expanded in the template
-	 * @return The expanded template.
-	 */
-	public String expandString(String s, Properties params) {
-		if(params == null || params.size() < 1) return s;
-
-		Pattern pattern = Pattern.compile( "\\{\\{\\s*\\S+\\s*\\}\\}" );
-		Matcher m = pattern.matcher(s);
-
-		StringBuffer sb = new StringBuffer();
-		 while (m.find()) {
-			 String raw = m.group();
-			 String varName = m.group().replace('{', ' ').replace('}', ' ').trim();
-			 String val = params.getProperty(varName);
-			 val = (val == null ? raw : val.trim());
-
-		     m.appendReplacement(sb, val);
-		 }
-		 m.appendTail(sb);
-
-		 return sb.toString();
-	}
+	
 
 	/**
 	 * Creates a handler.
